requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-admin-interface==0.19.0
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.0.1
python-slugify==6.1.1
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.9
